requirements.txt
streamlit
watchdog
openai
pypdf2
python-docx
python-dotenv
streamlit_js_eval
reportlab
markdown2